Keep track of first within the iterator
Previously, the iterator did not keep track of the first Node,
but just initialized a current node to the first node of the
RandomizedQueue class.  This unfortunately led to the Node running to the
end of the list, with no way of getting back to the front.

Accordingly, keep track of the first node in the iterator, and
reinitialized the current node to the first node at each calling
of the next method.