Don't null out current node until last
Previously, the next and previous attributes of the node being dequeued
were removed as the neighboring node's attributes were being repointed.
This led to deleting some of the information before it was needed, and
so to some null pointer exceptions.

Accordingly, set the attributes of the neighboring nodes first, then null
out the current node.