Copy existing RandomizedQueue into the iterator
Previously, the iterator referred to the objects in the RandomizedQueue.
This meant that when items were removed from the Iterator, they were also
removed from the queue.  This behavior was undesireable.

Accordingly, make new nodes containing the information and relationships
contained within the original RandomizedQueue structure, so that the
Iterator can change the relationships among these Nodes without mutating
the original Linked List.