Fix resizing errors
The previous commit attempted to address, in a roundabout way,
the ArrayIndexOutOfBoundsExceptions which were being generated
when the array was emptied and then enqueued to.  However,
it was not fully successful in that attempt, and it was
unclear why it should have been successful to begin with.

Previously, dequeue would check to see if the size of the new
array was at least 0, then would check to see if the size
equaled one quarter of the array's length, and if so it would
resize the array to one half of the array's length.
Unfortunately, this meant that when the new size was zero, the array
would be resized to an array of length 0, thus lacking even a
[0] element.  Accordingly, once the array had been emptied, it was
impossible to enqueue another item, because there was no place
for that new item to go.

Accordingly, don't resize the array unless the new size is
at least one.